pyproject.toml
[tool.poetry]
name = "django_site"
version = "0.1.0"
description = "Project description"
authors = ["GMSO <german.mene@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.6"
psycopg2-binary = "^2.9.1"
django-allauth = "^0.45.0"
whitenoise = "^5.3.0"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
django-extensions = "^3.1.5"
django-zen-queries = "^2.1.0"
django-compressor = "^2.4.1"
django-htmx = "^1.6.0"
pytest-cov = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.4.0"
black = "^21.8b0"
flake8 = "^3.9.2"
django-debug-toolbar = "^3.2.2"
ptvsd = "^4.3.2"
freezegun = "^1.1.0"
Werkzeug = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79